requirements.txt
gradio==3.50.2
soundfile==0.12.1
numpy>=1.21
requests
python-dotenv
openai
fastapi
pydantic
uvicorn